# account-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Account Service
  description: |
    **Microservicio de Gestión de Cuentas y Movimientos**

    <sup>(EN)</sup> This API handles bank account management and transaction movements.
    <sup>(ES)</sup> Esta API maneja la gestión de cuentas bancarias y movimientos transaccionales.

    **Arquitectura / Architecture:**
    - Spring Boot + WebFlux
    - Arquitectura Hexagonal / Hexagonal Architecture

    ---
    **Junior Level**
  contact:
    name: Brigeth Toapanta Andache
    email: margelybrigeth2001@gmail.com
  version: 1.1.0

servers:
  - url: http://localhost:8081/api/account-service
    description: ⁽ᴱᴺ⁾ Development Environment. ⁽ᴱˢ⁾ Entorno de Desarrollo.

tags:
  - name: Accounts
    description: ⁽ᴱᴺ⁾ Accounts CRUD operations. ⁽ᴱˢ⁾ Operaciones CRUD de cuentas.
  - name: Movements
    description: ⁽ᴱᴺ⁾ Transactional movement functions. ⁽ᴱˢ⁾ Funciones de movimientos transaccionales.

paths:
  /accounts:
    get:
      tags:
        - Accounts
      summary: ⁽ᴱᴺ⁾ Get accounts list. ⁽ᴱˢ⁾ Obtener listado de cuentas
      description: |
        ### Details:
        <sup>(EN)</sup> Returns a list of bank accounts. Filtering by customer and type.
        <sup>(ES)</sup> Retorna un listado de cuentas bancarias. Filtrado por cliente y tipo.
        ###
      operationId: getAccounts
      parameters:
        - name: customerId
          in: query
          description: ⁽ᴱᴺ⁾ Filter by client ID. ⁽ᴱˢ⁾ Filtrar por ID de cliente.
          required: false
          schema:
            $ref: '#/components/schemas/UUID'
        - name: accountType
          in: query
          description: ⁽ᴱᴺ⁾ Filter by account type. ⁽ᴱˢ⁾ Filtrar por tipo de cuenta.
          required: false
          schema:
            $ref: '#/components/schemas/AccountType'
      responses:
        '200':
          description: ⁽ᴱᴺ⁾ Account list retrieved successfully. ⁽ᴱˢ⁾ Listado de cuentas recuperado exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
              examples:
                SuccessfulResponse:
                  $ref: '#/components/examples/AccountsExample'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags:
        - Accounts
      summary: ⁽ᴱᴺ⁾ Create a new account. ⁽ᴱˢ⁾ Crear una nueva cuenta
      description: |
        ### Details:
        <sup>(EN)</sup> Create a new bank account associated with a client.
        <sup>(ES)</sup> Crear una nueva cuenta bancaria asociada a un cliente.
        
        - ⁽ᴱᴺ⁾ Client must exist and be active. ⁽ᴱˢ⁾ El cliente debe existir y estar activo.
        - ⁽ᴱᴺ⁾ Unique account number. ⁽ᴱˢ⁾ Número de cuenta único.
        - ⁽ᴱᴺ⁾ Initial balance cannot be negative. ⁽ᴱˢ⁾ El saldo inicial no puede ser negativo.
        ###
      operationId: createAccount
      requestBody:
        required: true
        description: ⁽ᴱᴺ⁾ Data for account creation. ⁽ᴱˢ⁾ Datos de la cuenta a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
            examples:
              AutoGeneratedNumber:
                summary: ⁽ᴱᴺ⁾ Auto-generated account number. ⁽ᴱˢ⁾ Número de cuenta auto-generado.
                $ref: '#/components/examples/AccountCreateExample'
              CustomAccountNumber:
                summary: ⁽ᴱᴺ⁾ Custom account number. ⁽ᴱˢ⁾ Número de cuenta personalizado.
                $ref: '#/components/examples/AccountCreateWithNumberExample'
      responses:
        '201':
          description: ⁽ᴱᴺ⁾ Account created successfully. ⁽ᴱˢ⁾ Cuenta creada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: ⁽ᴱᴺ⁾ Conflict - Account number already exists. ⁽ᴱˢ⁾ Conflicto - Número de cuenta ya existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: ⁽ᴱᴺ⁾ Get account by number. ⁽ᴱˢ⁾ Obtener por número de cuenta
      description: |
        ### Details:
        <sup>(EN)</sup> Returns complete data of a specific account.
        <sup>(ES)</sup> Retorna los datos completos de una cuenta específica.
        ###
      operationId: getAccountByNumber
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
      responses:
        '200':
          description: ⁽ᴱᴺ⁾ Account found successfully. ⁽ᴱˢ⁾ Cuenta encontrada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Accounts
      summary: ⁽ᴱᴺ⁾ Update account. ⁽ᴱˢ⁾ Actualizar cuenta
      description: |
        ### Details:
        <sup>(EN)</sup> Update account information (type and state).
        <sup>(ES)</sup> Actualizar información de la cuenta (tipo y estado).

        - ⁽ᴱᴺ⁾ Balance can only be modified through movements. ⁽ᴱˢ⁾ El saldo solo puede modificarse mediante movimientos.
        - ⁽ᴱᴺ⁾ Can reactivate inactive accounts. ⁽ᴱˢ⁾ Puede reactivar cuentas inactivas.
        ###
      operationId: updateAccount
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
      requestBody:
        required: true
        description: ⁽ᴱᴺ⁾ Data to update. ⁽ᴱˢ⁾ Datos a actualizar.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
            examples:
              UpdateAccountType:
                summary: ⁽ᴱᴺ⁾ Update account type. ⁽ᴱˢ⁾ Actualizar tipo de cuenta.
                value:
                  accountType: "CORRIENTE"
              DeactivateAccount:
                summary: ⁽ᴱᴺ⁾ Deactivate account. ⁽ᴱˢ⁾ Desactivar cuenta.
                value:
                  state: false
              ReactivateAccount:
                summary: ⁽ᴱᴺ⁾ Reactivate account. ⁽ᴱˢ⁾ Reactivar cuenta.
                value:
                  state: true
              UpdateBoth:
                summary: ⁽ᴱᴺ⁾ Update type and state. ⁽ᴱˢ⁾ Actualizar tipo y estado.
                $ref: '#/components/examples/AccountUpdateExample'
      responses:
        '200':
          description: ⁽ᴱᴺ⁾ Account updated successfully. ⁽ᴱˢ⁾ Cuenta actualizada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Accounts
      summary: ⁽ᴱᴺ⁾ Delete account. ⁽ᴱˢ⁾ Eliminar cuenta
      description: |
        ### Details:
        <sup>(EN)</sup> Performs a logical deletion of the account.
        <sup>(ES)</sup> Realiza una eliminación lógica de la cuenta.
        ###
      operationId: deleteAccount
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
      responses:
        '204':
          description: ⁽ᴱᴺ⁾ Account deleted successfully. ⁽ᴱˢ⁾ Cuenta eliminada exitosamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /movements:
    get:
      tags:
        - Movements
      summary: ⁽ᴱᴺ⁾ Get movements list. ⁽ᴱˢ⁾ Obtener listado de movimientos
      description: |
        ### Details:
        <sup>(EN)</sup> Returns a list of movements. Filtering by account and type.
        <sup>(ES)</sup> Retorna un listado de movimientos. Filtrar por cuenta y tipo.
        ###
      operationId: getMovements
      parameters:
        - name: accountNumber
          in: query
          description: ⁽ᴱᴺ⁾ Filter by account number. ⁽ᴱˢ⁾ Filtrar por número de cuenta.
          required: false
          schema:
            $ref: '#/components/schemas/AccountNumber'
        - name: movementType
          in: query
          description: ⁽ᴱᴺ⁾ Filter by movement type. ⁽ᴱˢ⁾ Filtrar por tipo de movimiento.
          required: false
          schema:
            $ref: '#/components/schemas/MovementType'
      responses:
        '200':
          description: ⁽ᴱᴺ⁾ Movement list retrieved successfully. ⁽ᴱˢ⁾ Listado de movimientos recuperado exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovementResponse'
              examples:
                SuccessfulResponse:
                  $ref: '#/components/examples/MovementsExample'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags:
        - Movements
      summary: ⁽ᴱᴺ⁾ Register new movement. ⁽ᴱˢ⁾ Registrar nuevo movimiento
      description: |
        ### Details:
        <sup>(EN)</sup> Create a new movement (debit or credit) in an account.
        <sup>(ES)</sup> Crear un nuevo movimiento (débito o crédito) en una cuenta.
        
        - ⁽ᴱᴺ⁾ Value must be greater than zero. ⁽ᴱˢ⁾ El valor debe ser mayor que cero.
        - ⁽ᴱᴺ⁾ Debit validation: sufficient balance required. ⁽ᴱˢ⁾ Validación de débito: se requiere saldo suficiente.
        - ⁽ᴱᴺ⁾ Atomic transactional record. ⁽ᴱˢ⁾ Registro transaccional atómico.
        ###
      operationId: createMovement
      requestBody:
        required: true
        description: ⁽ᴱᴺ⁾ Data for movement registration. ⁽ᴱˢ⁾ Datos del movimiento a registrar.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementCreateRequest'
            examples:
              CreditMovement:
                $ref: '#/components/examples/CreateCreditMovementExample'
              DebitMovement:
                $ref: '#/components/examples/CreateDebitMovementExample'
      responses:
        '201':
          description: ⁽ᴱᴺ⁾ Movement registered successfully. ⁽ᴱˢ⁾ Movimiento registrado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: ⁽ᴱᴺ⁾ Not Found - Account does not exist. ⁽ᴱˢ⁾ No encontrado - La cuenta no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ⁽ᴱᴺ⁾ Conflict - Insufficient funds for debit. ⁽ᴱˢ⁾ Conflicto - Saldo insuficiente para realizar el débito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: ⁽ᴱᴺ⁾ Unprocessable Entity - Account is inactive. ⁽ᴱˢ⁾ Entidad no procesable - La cuenta está inactiva.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    AccountNumberParam:
      name: accountNumber
      in: path
      description: ⁽ᴱᴺ⁾ Bank account number. ⁽ᴱˢ⁾ Número de cuenta bancaria.
      required: true
      schema:
        $ref: '#/components/schemas/AccountNumber'

  schemas:
    UUID:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"

    AccountNumber:
      type: integer
      format: int64
      description: ⁽ᴱᴺ⁾ Unique bank account number. ⁽ᴱˢ⁾ Número único de cuenta bancaria.
      minimum: 100000
      maximum: 9999999999
      example: 69845

    AccountType:
      type: string
      description: ⁽ᴱᴺ⁾ Bank account type. ⁽ᴱˢ⁾ Tipo de cuenta bancaria.
      enum: [AHORRO, CORRIENTE]
      example: "AHORRO"

    MovementType:
      type: string
      description: ⁽ᴱᴺ⁾ Transaction type (Credit/Debit). ⁽ᴱˢ⁾ Tipo de transacción (Crédito/Débito).
      enum: [DEBITO, CREDITO]
      example: "CREDITO"

    Money:
      type: number
      format: double
      description: ⁽ᴱᴺ⁾ Monetary value. ⁽ᴱˢ⁾ Valor monetario.
      minimum: 0
      multipleOf: 0.01
      example: 150.80

    PositiveMoney:
      type: number
      format: double
      description: ⁽ᴱᴺ⁾ Positive monetary value (greater than zero). ⁽ᴱˢ⁾ Valor monetario positivo (mayor que cero).
      exclusiveMinimum: true
      minimum: 0
      multipleOf: 0.01
      example: 100.00

    AccountResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
          description: ⁽ᴱᴺ⁾ Internal account ID. ⁽ᴱˢ⁾ ID interno de la cuenta.
        accountNumber:
          $ref: '#/components/schemas/AccountNumber'
        accountType:
          $ref: '#/components/schemas/AccountType'
        balance:
          $ref: '#/components/schemas/Money'
        state:
          type: boolean
          description: ⁽ᴱᴺ⁾ Account state (true=active, false=inactive). ⁽ᴱˢ⁾ Estado de la cuenta (true=activa, false=inactiva).
          example: true
        customerId:
          $ref: '#/components/schemas/UUID'
        customerName:
          type: string
          description: ⁽ᴱᴺ⁾ Client name (fetched from Customer Service). ⁽ᴱˢ⁾ Nombre del cliente (obtenido del servicio de clientes).
          example: "Maria Tupiza"

    AccountCreateRequest:
      type: object
      required: [accountType, customerId]
      properties:
        accountNumber:
          allOf:
            - $ref: '#/components/schemas/AccountNumber'
          description: ⁽ᴱᴺ⁾ Optional. If not provided, will be auto-generated. ⁽ᴱˢ⁾ Opcional. Si no se proporciona, se generará automáticamente.
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          allOf:
            - $ref: '#/components/schemas/Money'
          description: ⁽ᴱᴺ⁾ Initial balance. Default is 0. Cannot be negative. ⁽ᴱˢ⁾ Saldo inicial. Por defecto es 0. No puede ser negativo.
          default: 0.0
        customerId:
          $ref: '#/components/schemas/UUID'
        state:
          type: boolean
          description: ⁽ᴱᴺ⁾ Initial account state. ⁽ᴱˢ⁾ Estado inicial de la cuenta.
          default: true

    AccountUpdateRequest:
      type: object
      properties:
        accountType:
          $ref: '#/components/schemas/AccountType'
          description: ⁽ᴱᴺ⁾ Account type to update. ⁽ᴱˢ⁾ Tipo de cuenta a actualizar.
        state:
          type: boolean
          description: ⁽ᴱᴺ⁾ Account state (use to activate/deactivate). ⁽ᴱˢ⁾ Estado de la cuenta (usar para activar/desactivar).

    MovementResponse:
      type: object
      properties:
        movementId:
          $ref: '#/components/schemas/UUID'
        date:
          type: string
          format: date-time
          description: ⁽ᴱᴺ⁾ Movement registration date. ⁽ᴱˢ⁾ Fecha de registro del movimiento.
          example: "2025-11-06T10:30:00.000Z"
        movementType:
          $ref: '#/components/schemas/MovementType'
        amount:
          $ref: '#/components/schemas/PositiveMoney'
          description: ⁽ᴱᴺ⁾ Movement amount. ⁽ᴱˢ⁾ Monto del movimiento.
        balanceBefore:
          $ref: '#/components/schemas/Money'
          description: ⁽ᴱᴺ⁾ Account balance before movement. ⁽ᴱˢ⁾ Saldo de la cuenta antes del movimiento.
        balanceAfter:
          $ref: '#/components/schemas/Money'
          description: ⁽ᴱᴺ⁾ Account balance after movement. ⁽ᴱˢ⁾ Saldo de la cuenta después del movimiento.
        accountNumber:
          $ref: '#/components/schemas/AccountNumber'

    MovementCreateRequest:
      type: object
      required: [accountNumber, movementType, amount]
      properties:
        accountNumber:
          $ref: '#/components/schemas/AccountNumber'
        movementType:
          $ref: '#/components/schemas/MovementType'
        amount:
          $ref: '#/components/schemas/PositiveMoney'

    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          $ref: '#/components/schemas/UUID'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'

    ValidationError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: object

  responses:
    BadRequest:
      description: ⁽ᴱᴺ⁾ Bad Request - Validation Error. ⁽ᴱˢ⁾ Petición incorrecta - Error de validación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: ⁽ᴱᴺ⁾ Resource not found. ⁽ᴱˢ⁾ Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  examples:
    AccountsExample:
      value:
        - id: "a1b2c3d4-e5f6-4789-a012-3456789abcde"
          accountNumber: 76828
          accountType: "AHORRO"
          balance: 200.00
          state: true
          customerId: "550e8400-e29b-41d4-a716-446655440000"
          customerName: "Maria Tupiza"

    AccountCreateExample:
      value:
        accountType: "AHORRO"
        initialBalance: 200.00
        customerId: "550e8400-e29b-41d4-a716-446655440000"

    AccountCreateWithNumberExample:
      value:
        accountNumber: 478758
        accountType: "CORRIENTE"
        initialBalance: 500.00
        customerId: "550e8400-e29b-41d4-a716-446655440000"

    AccountUpdateExample:
      value:
        accountType: "CORRIENTE"
        state: false

    MovementsExample:
      value:
        - movementId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          date: "2025-11-06T10:30:00.000Z"
          movementType: "CREDITO"
          amount: 100.00
          balanceBefore: 200.00
          balanceAfter: 300.00
          accountNumber: 76828

    CreateCreditMovementExample:
      value:
        accountNumber: 478758
        movementType: "CREDITO"
        amount: 200.00

    CreateDebitMovementExample:
      value:
        accountNumber: 478758
        movementType: "DEBITO"
        amount: 100.00